Changed uint64_t enum to uint16_t enum

+++ src/MouseCodes.h
@@ -2,7 +2,7 @@
 
 #include <cstdint>
 
-enum class MouseCode : uint64_t
+enum class MouseCode : uint16_t
 {
 	// From glfw3.h
 	Button0 = 0,
